+++ libs/core/src/lib/combobox/combobox.component.ts
@@ -70,8 +70,8 @@ import { ComboboxMobileComponent } from './combobox-mobile/combobox-mobile.compo
     encapsulation: ViewEncapsulation.None,
     changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class ComboboxComponent implements ComboboxInterface, ControlValueAccessor, OnInit, OnChanges, AfterViewInit, OnDestroy {
-
+export class ComboboxComponent
+    implements ComboboxInterface, ControlValueAccessor, OnInit, OnChanges, AfterViewInit, OnDestroy {
     /** Values to be filtered in the search input. */
     @Input()
     dropdownValues: any[] = [];
@@ -145,6 +145,9 @@ export class ComboboxComponent implements ComboboxInterface, ControlValueAccesso
     /** Whether the matching string should be highlighted during filtration. */
     @Input()
     highlighting = true;
+
+    /** Whether the matching string should be highlighted after combobox value is selected. */
+    filterHighlight = true;
 
     /** Whether the popover should close when a user selects a result. */
     @Input()
@@ -241,9 +244,7 @@ export class ComboboxComponent implements ComboboxInterface, ControlValueAccesso
     ];
 
     /** Keys, that will close popover's body, when dispatched on search input */
-    readonly closingKeys: string[] = [
-        'Escape'
-    ];
+    readonly closingKeys: string[] = ['Escape'];
 
     /** Whether the combobox is opened. */
     open = false;
@@ -267,12 +268,10 @@ export class ComboboxComponent implements ComboboxInterface, ControlValueAccesso
     private readonly _onDestroy$: Subject<void> = new Subject<void>();
 
     /** @hidden */
-    onChange: any = () => {
-    };
+    onChange: any = () => {};
 
     /** @hidden */
-    onTouched: any = () => {
-    };
+    onTouched: any = () => {};
 
     /** @hidden */
     constructor(
@@ -280,8 +279,7 @@ export class ComboboxComponent implements ComboboxInterface, ControlValueAccesso
         private _menuKeyboardService: MenuKeyboardService,
         private _cdRef: ChangeDetectorRef,
         private _dynamicComponentService: DynamicComponentService
-    ) {
-    }
+    ) {}
 
     /** @hidden */
     ngOnInit(): void {
@@ -334,26 +332,25 @@ export class ComboboxComponent implements ComboboxInterface, ControlValueAccesso
         } else if (KeyUtil.isKey(event, this.closingKeys)) {
             this.isOpenChangeHandle(false);
             event.stopPropagation();
-        } else if (this.openOnKeyboardEvent &&
-            !event.ctrlKey &&
-            !KeyUtil.isKey(event, this.nonOpeningKeys)) {
+        } else if (this.openOnKeyboardEvent && !event.ctrlKey && !KeyUtil.isKey(event, this.nonOpeningKeys)) {
             this.isOpenChangeHandle(true);
         }
     }
 
     /** @hidden */
     onListKeydownHandler(event: KeyboardEvent): void {
-        const index: number = this.listItems.toArray().findIndex(
-            item => item.itemEl.nativeElement === document.activeElement
-        );
+        const index: number = this.listItems
+            .toArray()
+            .findIndex((item) => item.itemEl.nativeElement === document.activeElement);
         this._menuKeyboardService.keyDownHandler(event, index, this.listItems.toArray());
     }
 
     /** @hidden */
     onMenuClickHandler(value: any): void {
         if (value) {
-            const index: number = this.dropdownValues.findIndex(_value => _value === value);
+            const index: number = this.dropdownValues.findIndex((_value) => _value === value);
             this._handleClickActions(value);
+            this.filterHighlight = false;
             this.itemClicked.emit({ item: value, index: index });
         }
     }
@@ -422,6 +419,7 @@ export class ComboboxComponent implements ComboboxInterface, ControlValueAccesso
         this._resetDisplayedValues();
         this.isOpenChangeHandle(!this.open);
         this.searchInputElement.nativeElement.focus();
+        this.filterHighlight = false;
     }
 
     /** @hidden */
@@ -477,7 +475,7 @@ export class ComboboxComponent implements ComboboxInterface, ControlValueAccesso
     /** Method that picks other value moved from current one by offset, called only when combobox is closed */
     private _chooseOtherItem(offset: number): void {
         const activeValue: any = this._getOptionObjectByDisplayedValue(this.inputTextValue);
-        const index: number = this.dropdownValues.findIndex(value => value === activeValue);
+        const index: number = this.dropdownValues.findIndex((value) => value === activeValue);
         if (this.dropdownValues[index + offset]) {
             this.onMenuClickHandler(this.dropdownValues[index + offset]);
         }
@@ -494,8 +492,7 @@ export class ComboboxComponent implements ComboboxInterface, ControlValueAccesso
             .pipe(takeUntil(this._onDestroy$))
             .subscribe((index) => this.onMenuClickHandler(index));
         this._menuKeyboardService.focusEscapeBeforeList = () => this.searchInputElement.nativeElement.focus();
-        this._menuKeyboardService.focusEscapeAfterList = () => {
-        };
+        this._menuKeyboardService.focusEscapeAfterList = () => {};
     }
 
     /** @hidden */
@@ -515,6 +512,7 @@ export class ComboboxComponent implements ComboboxInterface, ControlValueAccesso
 
     /** @hidden */
     private _defaultFilter(contentArray: any[], searchTerm: any): any[] {
+        this.filterHighlight = true;
         if (typeof searchTerm === 'string') {
             const searchLower = searchTerm.toLocaleLowerCase();
             return contentArray.filter((item) => {
@@ -530,7 +528,6 @@ export class ComboboxComponent implements ComboboxInterface, ControlValueAccesso
             });
         }
     }
-
 
     /** @hidden*/
     private _popoverOpenHandle(): void {
@@ -584,7 +581,6 @@ export class ComboboxComponent implements ComboboxInterface, ControlValueAccesso
             this.displayedValues = this.dropdownValues;
         }
     }
-
 
     /** @hidden */
     private _hasDisplayedValues(): boolean {