+++ libs/platform/src/lib/components/form/public_api.ts
@@ -19,3 +19,4 @@ export * from './text-area/text-area.module';
 export * from './switch/switch.module';
 export * from './switch/switch/switch.config';
 export * from './switch/switch/switch.component';
+export * from './step-input/public_api';