+++ libs/platform/src/lib/components/form/step-input/public_api.ts
@@ -1,0 +1,4 @@
+export { StepInputConfig } from './step-input.config';
+export { StepInputAlign, StepInputStepFunctionAction, StepInputStepFunction } from './base.step-input';
+export { PlatformStepInputModule } from './step-input.module';
+export { NumberStepInputChangeEvent, NumberStepInputComponent } from './number/number-step-input.component';